pyproject.toml
[build-system]
requires = ["setuptools==70.0.0", "torch"]
build-backend = "setuptools.build_meta"

[project]
name = "MAST3R-SLAM"
version = "0.0.1"
dependencies = [
    "numpy>=1.26.4",
    "asmk",
    "scipy",
    "einops",
    "pyrealsense2",
    "evo",
    "natsort",
    # "torchcodec==0.1",
    "lietorch @ git+https://github.com/princeton-vl/lietorch.git",
    "plyfile",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mast3r_slam"]
namespaces = false
